+++ yash-semantics/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `yash-semantics` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - Unreleased
+## [0.4.0] - 2024-09-29
 
 ### Added
 
@@ -44,8 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       the `Exec` and `Interactive` shell options are both off.
 - External dependency versions:
     - Rust 1.77.0 → 1.79.0
-    - yash-env 0.2.0 → 0.3.0
-    - yash-syntax 0.10.0 → 0.11.0
+    - yash-env 0.2.0 → 0.4.0
+    - yash-syntax 0.10.0 → 0.12.0
 
 ## [0.3.0] - 2024-07-13